chore: move PairDisplayDiv to legacy
