+++ src/components/recorder/Pair.tsx
@@ -5,7 +5,7 @@ import { WorkflowFile } from "maxun-core";
 import { ClearButton } from "../ui/buttons/ClearButton";
 import { GenericModal } from "../ui/GenericModal";
 import { PairEditForm } from "./PairEditForm";
-import { PairDisplayDiv } from "./PairDisplayDiv";
+import { PairDisplayDiv } from "../../../legacy/src/PairDisplayDiv";
 import { EditButton } from "../ui/buttons/EditButton";
 import { BreakpointButton } from "../ui/buttons/BreakpointButton";
 import VisibilityIcon from '@mui/icons-material/Visibility';
